requirements.txt
requests==2.32.3
Jinja2==3.1.4
PyYAML==6.0.2


